Reopen PR #86

**Description**: What this PR does
ArrayEach() correctly parse empty array, or with whitespace (#79)
Change golang version on travis ci to 1.7 (need for testing method `t.Run()`)

**Benchmark before change**:
```
BenchmarkJsonParserLarge-4                    	  100000	     66302 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	 1000000	     10059 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      6535 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      6987 ns/op	     656 B/op	      13 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	 1000000	     10135 ns/op	     608 B/op	      12 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	       885 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       695 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	      1000 ns/op	     256 B/op	       9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       857 ns/op	     240 B/op	       8 allocs/op
```

**Benchmark after change**:
```
BenchmarkJsonParserLarge-4                    	  100000	     65816 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	 1000000	     10050 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      6591 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      7258 ns/op	     656 B/op	      13 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	 1000000	     10519 ns/op	     608 B/op	      12 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	       889 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       688 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	       993 ns/op	     256 B/op	       9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       863 ns/op	     240 B/op	       8 allocs/op
```